// src/interfaces/common.ts
import { USER_ROLE } from '../enums/user';
import { IGenericErrorMessage } from './error';

export type IGenericResponse<T> = {
  meta: {
    page: number;
    limit: number;
    total: number;
  };
  data: T;
};

export type IGenericErrorResponse = {
  statusCode: number;
  message: string;
  errorMessages: IGenericErrorMessage[];
};


export type IAuthUser = {
  userId: string;
  role: USER_ROLE,
  email: string
} | null